requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3

